pyproject.toml
[project]
name = "marketintel"
version = "0.1.0"
description = "Market intelligence data platform with Django, Celery, and Dagster"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Humberto Henriquez" }]

dependencies = [
    # Web framework
    "django>=5.0,<5.1",
    "djangorestframework>=3.15",
    "django-filter>=24.2",
    "drf-spectacular>=0.27",
    "django-cors-headers>=4.3",
    # Database
    "psycopg[binary]>=3.1",
    "django-redis>=5.4",
    # Async / distributed
    "celery>=5.3",
    "redis>=5.0",
    "kombu>=5.3",
    # Orchestration
    "dagster>=1.7",
    "dagster-webserver>=1.7",
    "dagster-postgres>=0.23",
    "dagster-celery>=0.23",
    # HTTP & validation
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Storage
    "boto3>=1.34",
    "duckdb>=0.10",
    # Observability
    "structlog>=24.1",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-django>=0.46b0",
    "opentelemetry-instrumentation-celery>=0.46b0",
    "django-prometheus>=2.3",
    # Production server
    "gunicorn>=22.0",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "vcrpy>=6.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "django-stubs[compatible-mypy]>=5.0",
    "ipython>=8.24",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations", "*.pyi"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "DJ", "PT", "RUF"]
ignore = ["E501"]  # handled by formatter

[tool.ruff.lint.per-file-ignores]
"**/settings/*.py" = ["F403", "F405"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra --strict-markers --cov=apps --cov=pipelines --cov-report=term-missing"
markers = [
    "unit: fast unit tests, no DB",
    "integration: requires DB / Redis",
    "vcr: uses recorded HTTP fixtures",
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
strict = true
warn_return_any = false
disallow_untyped_decorators = false

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apps*", "config*", "pipelines*", "warehouse*"]
